Protect routes

+++ products/views.py
@@ -1,4 +1,5 @@
 from django.shortcuts import render, get_object_or_404, redirect
+
 from .models import Product
 from .forms import AddProductForm
 
@@ -21,36 +22,45 @@ def product_details(request, pk):
 
 
 def product_add(request):
-    if request.method == 'POST':
-        form = AddProductForm(request.POST, request.FILES)
+    if request.user.is_authenticated and request.user.is_superuser:
+        if request.method == 'POST':
+            form = AddProductForm(request.POST, request.FILES)
 
-        if form.is_valid():
-            form.save()
-            return render(request, 'products/product-add-success.html')
+            if form.is_valid():
+                form.save()
+                return render(request, 'products/product-add-success.html')
 
+        else:
+            form = AddProductForm()
+
+        return render(request, 'products/product-add.html', {'form': form})
     else:
-        form = AddProductForm()
-
-    return render(request, 'products/product-add.html', {'form': form})
+        return redirect('products_list')
 
 
 def product_edit(request, pk):
-    product = get_object_or_404(Product, pk=pk)
+    if request.user.is_authenticated and request.user.is_superuser:
+        product = get_object_or_404(Product, pk=pk)
 
-    if request.method == 'POST':
-        form = AddProductForm(request.POST, request.FILES, instance=product)
+        if request.method == 'POST':
+            form = AddProductForm(
+                request.POST, request.FILES, instance=product)
 
-        if form.is_valid():
-            form.save()
-            return render(request, 'products/product-add-success.html', {'edit': True})
+            if form.is_valid():
+                form.save()
+                return render(request, 'products/product-add-success.html', {'edit': True})
 
+        else:
+            form = AddProductForm(instance=product)
+
+        return render(request, 'products/product-add.html', {'form': form, 'edit': True})
     else:
-        form = AddProductForm(instance=product)
-
-    return render(request, 'products/product-add.html', {'form': form, 'edit': True})
+        return redirect('products_list')
 
 
 def product_delete(request, pk):
-    product = Product.objects.get(pk=pk)
-    product.delete()
+    if request.user.is_authenticated and request.user.is_superuser:
+        product = Product.objects.get(pk=pk)
+        product.delete()
+
     return redirect("products_list")
